Fixed score display, added pause menu, updated pixel sampler

+++ ASCII_GAME/SpriteAnimation.h
@@ -9,7 +9,7 @@ public:
 	~SpriteAnimation();
 
 	Vec2<float> GetPosition() const { return m_Position; }
-	Vec2<int> GetSize() const { return m_Size; }
+	Vec2<short> GetSize() const { return m_Size; }
 
 	void SetSpriteSheet(const char* tgaFilePath, int frameCount);
 	void SetFrame(int frameNumber) { m_CurrentFrame = frameNumber; }
@@ -31,5 +31,5 @@ private:
 	bool m_bActive;
 
 	Vec2<float> m_Position;
-	Vec2<int> m_Size;
+	Vec2<short> m_Size;
 };
